Se realizan cambios en los modelos y se crea Config.
Se actualizan las propiedades de los modelos básicos de la aplicación. Se crean el modulo de configuración así como un componente Configuration para ello.

+++ MassiveSsh/Models/Kvr.cs
@@ -30,7 +30,7 @@ namespace Acabus.Models
         /// </summary>
         /// <param name="id">Identificador del kiosko.</param>
         /// <param name="station">Estación a la que pertenece.</param>
-        public Kvr(ushort id, Station station, String numeSeri) : base(id, DeviceType.KVR, station, numeSeri) { }
+        public Kvr(UInt64 id, Station station, String numeSeri) : base(id, DeviceType.KVR, station, numeSeri) { }
 
         /// <summary>
         /// Crea una instancia básica de KVR.
